chore: fix pre-commit config and adapt code to make all checks pass (#162)
* chore: fix pre-commit versions and clang-format hook failure

* chore(docs): fix documentation pre-commit errors

* chore: fix pre-commit and compiler warnings in all code files modified between main..develop

* chore(yamllint): ignore yaml document start in clang-format config as it is required

* chore: add copyright and license headers

* chore(pre-commit): allow multiple documents in YAML files to make .clang-format file check pass

* chore: make pre-commit pass for parameter schema code

* chore: add copyright and license to all source files

* chore: implement pre-commit suggestions in all CPP/HPP files

* chore: whitespace changes in non-cpp files to satisfy pre-commit

* chore: flake8 changes to satisfy pre-commit

* fix: allow implicit conversion to status types again

* chore: clean up imports

* chore: add override/inline where necessary

* chore(nebula_ros): remove obsolete wrapper base types

* chore: move nolint comments to be robust to formatting linebreaks

* chore(velodyne): fix indentation in velodyne calibration files to satisfy yamllint and ignore them in prettier to prevent conflicting styles

* chore(hesai): fix decoder test config yaml quotations

* chore: whitespace changes

* chore: remove empty, un-parseable local.cspell.json

* chore(prettier): ignore yaml and json files as they are handled by yamllint/clang-tidy respectively

* chore: make pre-commit pass on new files after #146

* chore: update cspell to pass spell-check

* chore: rename contributing.md docs page to make failing link check pass

+++ nebula_common/include/nebula_common/robosense/robosense_common.hpp
@@ -1,14 +1,27 @@
+// Copyright 2024 TIER IV, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 #pragma once
 
 #include "nebula_common/nebula_common.hpp"
 #include "nebula_common/nebula_status.hpp"
 
-#include <bitset>
 #include <cmath>
 #include <fstream>
 #include <iostream>
-#include <optional>
 #include <sstream>
+#include <string>
 #include <vector>
 
 namespace nebula
@@ -74,7 +87,7 @@ struct RobosenseCalibrationConfiguration : CalibrationConfigurationBase
     std::getline(stream, header);
 
     char sep;
-    int laser_id;
+    size_t laser_id;
     float elevation;
     float azimuth;
     Status load_status = Status::OK;
@@ -174,10 +187,10 @@ struct RobosenseCalibrationConfiguration : CalibrationConfigurationBase
 
   void CreateCorrectedChannels()
   {
-    for(auto& correction : calibration) {
+    for (auto & correction : calibration) {
       uint16_t channel = 0;
-      for(const auto& compare:calibration) {
-        if(compare.elevation < correction.elevation) ++channel;
+      for (const auto & compare : calibration) {
+        if (compare.elevation < correction.elevation) ++channel;
       }
       correction.channel = channel;
     }
